Don't fail when view has been already compiled by ActionView. Fixes #74
We rely on ActionView to retrieve the source of files when calling `extends`.
But during the process of rendering templates, ActionView compiles them into a
method and discard the original source code, leading to the bug described in

Now, if the source of the ActionView template is nil, i.e. the template has
already been rendered, we call the `refresh` method on the template which is
going to fetch the source code from the disk while not busting the compilation
already done by ActionView.

This bug only happened when the option `cache_templates` is set to false, or in
development environment by default.